WordPress Updates

+++ wp-content/plugins/supreme-modules-for-divi/uninstall.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Fired when the plugin is uninstalled.
  *
@@ -30,8 +29,9 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
-if ( get_option('dsm_settings_misc')['dsm_uninstall_on_delete'] == 'on' ) {
+if ( isset( get_option( 'dsm_settings_misc' )['dsm_uninstall_on_delete'] ) && get_option( 'dsm_settings_misc' )['dsm_uninstall_on_delete'] === 'on' ) {
 	delete_option( 'dsm_general' );
+	delete_option( 'dsm_settings_social_media' );
 	delete_option( 'dsm_settings_misc' );
 	delete_option( 'dsm-supreme-modules-for-divi-activation-date' );
 }
